Title: add wifi udp
Comment:
1.  Remove socket management from WiFiClass. And use lwip's socket management.
2.  To check if any packet comes in, use lwip's peek instead of read 1 byte and buffer it.
3.  Add 2 wifi udp example: WiFiUdpNtpClient and WiFiUdpSendReceiveString
4.  Exchange order of header between arduino and ameba

+++ hardware/libraries/WiFi/src/WiFiServer.h
@@ -9,11 +9,13 @@ class WiFiClient;
 class WiFiServer : public Server {
 private:
   	uint16_t _port;
-  	uint8_t _sock_ser;
+  	int _sock_ser;
 	ServerDrv serverfd;
 public:
   	WiFiServer(uint16_t);
+
   	WiFiClient available(uint8_t* status = NULL);
+
   	void begin();
   	virtual size_t write(uint8_t b);
   	virtual size_t write(const uint8_t *buf, size_t size);
